+++ e2e/zsh/test_zsh_pty.py
@@ -11,7 +11,7 @@ from conftest import set_working_dir
 
 
 def get_ansi_bracketed_pastes(complgen_binary_path: Path, grammar: str, input: bytes, working_dir: Path | None = None) -> list[str]:
-    completion_script = subprocess.run([complgen_binary_path, 'compile', '--zsh-script', '-', '-'], input=grammar.encode(), stdout=subprocess.PIPE, stderr=sys.stderr, check=True).stdout.decode()
+    completion_script = subprocess.run([complgen_binary_path, 'aot', '--zsh-script', '-', '-'], input=grammar.encode(), stdout=subprocess.PIPE, stderr=sys.stderr, check=True).stdout.decode()
     (pid, fd) = pty.fork()
     if pid == 0:
         # We're in the child